potsofgold.cpp
The players can see how many coins are there in each gold pot, and each player gets alternating turns in which the player can pick a pot from either end of the line. The winner is the player who has a more number of coins at the end compared to the other player.